Delete Confirm Alert

+++ frontend/src/components/GoalItem.jsx
@@ -1,12 +1,28 @@
 import { useDispatch } from "react-redux";
+import { useRef } from "react";
 import { updateGoal, deleteGoal } from "../features/goals/goalSlice";
 
 import { FaTrashAlt } from "react-icons/fa";
 import moment from "moment";
 
-import { Checkbox, Button, Heading, Text } from "@chakra-ui/react";
+import {
+	Checkbox,
+	Button,
+	Heading,
+	Text,
+	useDisclosure,
+	AlertDialog,
+	AlertDialogBody,
+	AlertDialogFooter,
+	AlertDialogHeader,
+	AlertDialogContent,
+	AlertDialogOverlay,
+} from "@chakra-ui/react";
 
 function GoalItem({ goal }) {
+	const { isOpen, onOpen, onClose } = useDisclosure();
+	const cancelRef = useRef();
+
 	const dispatch = useDispatch();
 
 	const getDate = (date) => {
@@ -47,11 +63,43 @@ function GoalItem({ goal }) {
 				leftIcon={<FaTrashAlt />}
 				colorScheme="red"
 				variant="solid"
-				onClick={() => dispatch(deleteGoal(goal._id))}
+				onClick={onOpen}
 				width="20%"
 			>
 				Delete
 			</Button>
+
+			<AlertDialog
+				isOpen={isOpen}
+				leastDestructiveRef={cancelRef}
+				onClose={onClose}
+				isCentered
+			>
+				<AlertDialogOverlay>
+					<AlertDialogContent>
+						<AlertDialogHeader fontSize="lg" fontWeight="bold">
+							Delete Goal
+						</AlertDialogHeader>
+
+						<AlertDialogBody>
+							Are you sure? You can't undo this action afterwards.
+						</AlertDialogBody>
+
+						<AlertDialogFooter>
+							<Button ref={cancelRef} onClick={onClose}>
+								Cancel
+							</Button>
+							<Button
+								colorScheme="red"
+								onClick={() => dispatch(deleteGoal(goal._id))}
+								ml={3}
+							>
+								Delete
+							</Button>
+						</AlertDialogFooter>
+					</AlertDialogContent>
+				</AlertDialogOverlay>
+			</AlertDialog>
 		</div>
 	);
 }
